% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainyRASCH.R, R/thresholds.R
\name{rainyRASCHgrp}
\alias{rainyRASCHgrp}
\alias{beta2delta}
\alias{delta2beta}
\title{Transformation of beta parameters to delta (item-category threshold) parameters and vice versa}
\usage{
rainyRASCHgrp(delta)

beta2delta(beta, x)

delta2beta(delta, x)
}
\arguments{
\item{delta}{Matrix with K x M item-category threshold parameters for K items with at most M categories.}

\item{beta}{Matrix (K x M) or vector (K*M) of beta parameters for K items with at most M categories.}

\item{x}{Matrix (nobs x K) of K item responses. Only used if \code{beta} is a vector to transform \code{beta} to a matrix.}
}
\value{
Matrix of item parameters.
}
\description{
\code{beta2delta} transforms the \eqn{\beta}{beta} parameters in \deqn{P(X_{vi} = x) = \frac{\exp(x\theta_v + \beta_{ix})}{\sum_{l=0}^{m_i}\exp(l\theta_v + \beta_{il})}}{P(X_{vi} = x) = (exp(x*theta_v + beta_{ix}))/(sum_{l=0}^{m_i} exp(l*theta_v + beta_{il}))} into delta (item-category threshold) parameters.
\code{delta2beta} transforms the \eqn{delta}{delta} (item-category threshold) parameters in \deqn{P(X_{vi} = x) = \frac{\exp(x\theta_v - \sum_{h=1}^x\delta_{il})}{\sum_{l=0}^{m_i}\exp(l\theta_v - \sum_{h=l}^x\delta_{il})}}{P(X_{vi} = x) = (exp(x*theta_v - sum_{h=1}^x delta_{il}))/(sum_{l=0}^{m_i} exp(l*theta_v - sum_{h=l}^x delta_{il}))} into delta (item-category threshold) parameters.
}
\section{Functions}{
\itemize{
\item \code{rainyRASCHgrp()}: Transformation of delta (item-category threshold) parameters to beta parameters

\item \code{delta2beta()}: Transformation of delta (item-category threshold) parameters to beta parameters

}}
\examples{
library(eRm)
data(SPADI)
SPADI.complete <- SPADI[complete.cases(SPADI), ]
it.SPADI <- SPADI.complete[, 9:16]
model.SPADI <- eRm::PCM(it.SPADI)
beta <- model.SPADI$betapar

betamat <- beta2mat(x = it.SPADI, beta = model.SPADI$betapar)
delta <- beta2delta(beta = betamat)
delta.eRm <- eRm::thresholds(model.SPADI)

betaTjek <- delta2beta(delta = delta) # Tjek mod betamat


}
\references{
Mair, P., & Hatzinger, R. . (2007). Extended Rasch Modeling: The eRm Package for the Application of IRT Models in R. Journal of Statistical Software, 20(9), 1–20. https://doi.org/10.18637/jss.v020.i09
}
