% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta2mat.R
\name{beta2mat}
\alias{beta2mat}
\title{Parameter vector to parameter matrix function}
\usage{
beta2mat(x, beta)
}
\arguments{
\item{x}{Matrix (nobs x K) of K item responses.}

\item{beta}{Vector of item parameters.}
}
\value{
Matrix of item parameters.
}
\description{
This function transform a vector of item parameters from a polytomous Rasch model to a matrix of item parameters where rows correspond to item categories and columns correspond to items or vice versa.
}
\details{
...
}
\examples{
library(eRm)
data(SPADI)
SPADI.complete <- SPADI[complete.cases(SPADI), ]
it.SPADI <- SPADI.complete[, 9:16]
model.SPADI <- eRm::PCM(it.SPADI)
betamat <- beta2mat(x = it.SPADI, beta = model.SPADI$betapar)

}
