% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bystat.RASCHstats.R
\name{bystat}
\alias{bystat}
\alias{bystat.RASCHstats}
\title{Print function}
\usage{
bystat(x)

\method{bystat}{RASCHstats}(x)
}
\arguments{
\item{x}{A \code{RASCHstats} object.}
}
\description{
Print list (by type of fit statistics) of matrices of sampled statistics for a \code{RASCHstats} object.
}
