% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CICCplot.R, R/CICCplotpackage.R
\docType{package}
\name{CICCplot}
\alias{CICCplot}
\title{Create CICC plot}
\usage{
CICCplot(
  model,
  which.item = 1,
  lower.groups = NULL,
  all.items = FALSE,
  grid.items = TRUE,
  error.bar = TRUE,
  col.cicc = NULL,
  col.error.bar = NULL
)
}
\arguments{
\item{model}{Object of class `Rm`.}

\item{which.item}{Item to plot. Default is 1.}

\item{lower.groups}{A vector used for grouping the set of total scores into intervals. The vector should contain the lower points of the intervals, that the set of possible total scores should be divided into. If zeroincluded, it will be added automatically. It is also possible to state `lower.groups="all"` (default) to have empirical expected item-scores for every possible total score in the figure.}

\item{all.items}{Logical flag for constructing CICC plots for all items in the data. Default is `all.items = FALSE`.}

\item{grid.items}{Grid}

\item{error.bar}{Default is `error.bar = TRUE`.}

\item{col.cicc}{The color of the CICC. NULL by default, in which case the CICC is black.}

\item{col.error.bar}{The color of the error bars. NULL by default, in which case the error bars are red.}
}
\value{
CICC plot
}
\description{
Description...
}
\examples{
library(iarm)
str(amts)
it.AMTS <- amts[, 4:13]
it.AMTS.complete <- it.AMTS[complete.cases(it.AMTS), ]
model.AMTS <- RM(it.AMTS.complete, sum0 = FALSE)
CICCplot(model = model.AMTS, lower.groups = "all")
CICCplot(model = model.AMTS, lower.groups = "all", error.bar = FALSE)

}
\author{
Ann-Sophie Buchardt \email{asbu@sund.ku.dk}
}
