% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudRM.R
\name{cloud}
\alias{cloud}
\title{Simulate Data under a Rasch model}
\usage{
cloud(
  items,
  group,
  method.item = c("PCML", "CML", "JML", "MML"),
  B,
  model = c("RMD", "RMP"),
  trace.it = 0
)
}
\arguments{
\item{items}{items}

\item{group}{group}

\item{method.item}{Estimation method for item parameters.}

\item{B}{Number of simulations.}

\item{model}{Character. "RMD" for dichotomous Rasch model. "RMP" for partial credit model.}

\item{trace.it}{If \code{trace.it=1}, then progress bars are displayed.}
}
\description{
Repeated random sampling to obtain item responses from estimated item and person parameters.
}
\examples{
data(SPADI)
SPADI.complete <- SPADI[complete.cases(SPADI), ]
group <- as.factor(SPADI.complete$gender)
SPADI1 <- SPADI.complete[group == 1,]
SPADI2 <- SPADI.complete[group == 2,]
it.SPADI1 <- SPADI1[, 9:16]
it.SPADI2 <- SPADI2[, 9:16]
items = SPADI.complete
B = 10

}
