% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CUTscores.R
\name{CUTscores}
\alias{CUTscores}
\title{Grouping of total scores function}
\usage{
CUTscores(x, probs = seq(0, 1, length.out = 6), min0 = TRUE)
}
\arguments{
\item{x}{A numeric vector which is to be converted to a factor by cutting.}

\item{probs}{A numeric vector of probabilities with values in [0,1].}

\item{min0}{Logical flag for shifting responses such that lowest category is 0. Default is \code{TRUE}.}
}
\value{
Vector
}
\description{
This function groups total scores by quantiles.
}
\examples{
library(eRm)
data(SPADI)
SPADI.complete <- SPADI[complete.cases(SPADI), ]
it.SPADI <- SPADI.complete[, 9:16]
model.SPADI <- eRm::PCM(it.SPADI)
lower.groups <- CUTscores(x = it.SPADI, probs = seq(0, 1, length.out = 6))
CICCplot(model = model.SPADI, lower.groups = lower.groups)

}
