% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIFplot.R
\name{DIFplot}
\alias{DIFplot}
\title{Stratified Conditional Item Characteristic Curves}
\usage{
DIFplot(
  model,
  which.item = 1,
  strat.vars,
  lower.groups = "all",
  grid.items = FALSE,
  error.bar = TRUE,
  error.level = 0.95,
  dodge.width = 0.5,
  group.connect = FALSE,
  point.size = 1,
  line.size = 1,
  line.type = 1,
  errorbar.width = 0,
  errorbar.size = 1,
  error.band = FALSE,
  error.band.col = "gray",
  smooth.error.band = FALSE,
  ...
)
}
\arguments{
\item{model}{A model object of class \code{Rm} or \code{eRm} returned from the functions \code{RM()}, \code{PCM()}, or \code{RSM()} from the \code{eRm} package.}

\item{which.item}{An indexing vector specifying the item(s) for which a CICC-plot should be constructed. Either an integer or vector of integers to select items by position, or a character string or a vector of character strings to select items by their names. Or a character string \code{"all"} for constructing CICC plots for all items in the data. The default is \code{which.item = 1}.}

\item{strat.vars}{A named list of categorical variables for stratification.}

\item{lower.groups}{A named list of length \code{length(strat.vars)} of lists, vectors or a single vector for grouping the set of possible total scores into intervals, for which the empirical expected item-score will be calculated and added to the plot. The vector(s) should contain the lower points of the intervals, that the set of possible total scores should be divided into. If zero does not appear in the vector(s), it will be added automatically. If \code{lower.groups = "all"} (default), the empirical expected item-score will be plotted for every possible total score. If a list is provided, the arguments should be named corresponding to the \code{strat.vars}. If the arguments are lists, they should be named corresponding to the levels of the \code{strat.vars}.}

\item{grid.items}{Logical flag for arranging the items selected by which.item in grids using the \code{ggarrange} function from the \code{ggpubr} package. Default value is \code{FALSE}.}

\item{error.bar}{Logical flag for adding error bars illustrating the empirical confidence interval of the observed means of the conditional item score. The confidence intervals are calculated as follows: For each interval l of the total score, induced by the lower-groups argument, the mean x_l, variance var(x_l), and number of observations n_l within the interval of the total score will be calculated. The confidence interval for the mean x_l is then found as \eqn{x_l \pm 2\cdot \sqrt(\frac{var(x_l)}{n_l})}. Default value is \code{TRUE}.}

\item{error.level}{The confidence level required. Default is a 95% confidence level.}

\item{dodge.width}{Dodging width of error bars. To prevent overlapping error bars, dodging (jittering) preserves the vertical position of error bars while adjusting the horizontal position. Default is \code{dodge.width = 0.5}.}

\item{group.connect}{Logical flag for connecting grouped total scores. Default value is \code{FALSE}.}

\item{point.size}{Size aesthetics for \code{geom_point()}.}

\item{line.size}{Size aesthetics for \code{geom_line()}.}

\item{line.type}{Linetype aesthetics for \code{geom_line()}.}

\item{errorbar.width}{Width aesthetics for \code{geom_line()}.}

\item{errorbar.size}{Size aesthetics for \code{geom_errorbar()}.}

\item{error.band}{Default is FALSE.}

\item{error.band.col}{Colour of error band (if \code{error.band = TRUE}).}

\item{smooth.error.band}{Logical flag.}

\item{...}{Arguments to be passed to \code{ggarrange}. The arguments will only be used if \code{grid.items = TRUE}.}
}
\value{
CICC plot
}
\description{
This function constructs Conditional Item Characteristic Curves for selected items in a Rasch-model and allows for stratification on categorical variables. These plots can be used to investigate item misfit and differential item functioning (DIF).
}
\details{
...
}
\examples{
library(eRm)
data(SPADI)
it.SPADI <- SPADI[, 9:16]
model.SPADI <- eRm::PCM(it.SPADI)
strat.vars <- list(gender = SPADI[, "gender"])
DIFplot(model = model.SPADI, strat.vars = strat.vars)
strat.vars <- list(gender = SPADI[, "gender"], over60 = SPADI[, "over60"])
DIFplot(model = model.SPADI, strat.vars = strat.vars)
DIFplot(model = model.SPADI, which.item = c(1,2), strat.vars = strat.vars)
lower.groups <- list(gender = list("1" = c(0, 5, 15, 20, 30, 40),
                                   "2" = c(0,  5, 10)),
                     over60 = list("0" = c(0, 10, 15, 20),
                                   "1" = c(0, 8, 15, 17, 18, 22, 29, 38)))
DIFplot(model = model.SPADI, strat.vars = strat.vars, lower.groups = lower.groups)
lower.groups <- list(gender = list("1" = c(0, 5, 15, 20, 30, 40),
                                   "2" = c(0,  5, 10)),
                     over60 = list("0" = c(0, 5, 15, 20, 30, 40),
                                   "1" = c(0,  5, 10)))
DIFplot(model = model.SPADI, strat.vars = strat.vars, lower.groups = lower.groups)

DIFplot(model = model.SPADI, strat.vars = strat.vars,
        lower.groups = lower.groups, group.connect = TRUE)

DIFplot(model = model.SPADI, strat.vars = strat.vars,
        lower.groups = lower.groups, group.connect = TRUE,
        error.band = TRUE, smooth.error.band = TRUE)
}
