% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPImap_hist.R
\name{.reconcileScatPlot}
\alias{.reconcileScatPlot}
\title{Internal ggPImap functions}
\usage{
.reconcileScatPlot(p, beta, threshtable)
}
\arguments{
\item{p}{p}

\item{beta}{beta}

\item{threshtable}{threshtable}
}
\description{
Internal ggPImap functions
}
