% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPImap.R
\name{ggPImap}
\alias{ggPImap}
\title{Histogram for Person-Item Map}
\usage{
ggPImap(
  p,
  theta,
  delta,
  group = NULL,
  type = c("histogram", "density", "boxplot", "violin", "densigram"),
  margins = c("top", "right", "bottom", "left"),
  size = 5,
  ggtheme = theme_void,
  ...,
  xparams = list(),
  yparams = list(),
  groupColour = FALSE,
  groupFill = FALSE
)
}
\arguments{
\item{p}{A \link[RASCHplot]{ggImap} to add marginal plots to. If \code{p} is
not provided, then all of \code{theta}, and \code{delta} must be
provided.}

\item{theta}{The person parameters to use for creating the marginal plots.}

\item{delta}{Input delta (item-category threshold) parameters. This must be a matrix or list of matrices in case of stratification. Ignored if \code{object} is provided. Ignored if \code{p} is provided.}

\item{group}{group}

\item{type}{What type of marginal plot to show. One of: [density, histogram, boxplot, violin, densigram]
(a "densigram" is when a density plot is overlaid on a histogram).}

\item{margins}{Along which margins to show the plots. One of: [both, x, y].}

\item{size}{Integer describing the relative size of the marginal plots
compared to the main plot. A size of 5 means that the main plot is 5x wider
and 5x taller than the marginal plots.}

\item{ggtheme}{ggplot2 function or theme object. Default value is \code{theme_void}. Allowed values are the official ggplot2 themes including \code{theme_gray}, \code{theme_bw}, \code{theme_classic}, \code{theme_void}, .... Theme objects are also allowed (e.g., \code{theme_classic()}).}

\item{...}{Extra parameters to pass to the marginal plots. Any parameter that
\code{geom_line()}, \code{geom_histogram()}, \code{geom_boxplot()}, or \code{geom_violin()} accepts
can be used. For example, \code{colour = "red"} can be used for any marginal plot type,
and \code{binwidth = 10} can be used for histograms.}

\item{xparams}{List of extra parameters to use only for the marginal plot along
the x axis.}

\item{yparams}{List of extra parameters to use only for the marginal plot along
the y axis.}

\item{groupColour}{If \code{TRUE}, the colour (or outline) of the marginal
plots will be grouped according to the variable mapped to \code{colour} in the
scatter plot. The variable mapped to \code{colour} in the scatter plot must
be a character or factor variable. See examples below.}

\item{groupFill}{If \code{TRUE}, the fill of the marginal
plots will be grouped according to the variable mapped to \code{colour} in the
scatter plot. The variable mapped to \code{colour} in the scatter plot must
be a character or factor variable. See examples below.}
}
\description{
Visualise the alignment of the item and person parameters.
}
\examples{
data(SPADI)
SPADI.complete <- SPADI[complete.cases(SPADI), ]
it.SPADI <- SPADI.complete[, 9:16]
object <- eRm::PCM(it.SPADI)
pp <- eRm::person.parameter(object)
p <- ggImap(object)
ggPImap(p, theta = unlist(pp$thetapar))

delta <- beta2delta(beta = object$betapar, x = it.SPADI)
p <- ggImap(delta = delta)
ggPImap(p, theta = unlist(pp$thetapar))

ggPImap(p, theta = unlist(pp$thetapar), group = as.factor(SPADI.complete$gender[-pp$pers.ex]))

if (FALSE) {
SPADI1 <- SPADI[SPADI$gender == 1,]
SPADI2 <- SPADI[SPADI$gender == 2,]
SPADI.complet1 <- SPADI1[complete.cases(SPADI1), ]
SPADI.complet2 <- SPADI2[complete.cases(SPADI2), ]
it.SPADI1 <- SPADI.complet1[SPADI.complet1$gender == 1, 9:16]
it.SPADI2 <- SPADI.complet2[SPADI.complet2$gender == 2, 9:16]
object1 <- eRm::PCM(it.SPADI1)
object2 <- eRm::PCM(it.SPADI2)
beta1 <- object1$betapar
beta2 <- object2$betapar
pp1 <- eRm::person.parameter(object2) #eRm::person.parameter(object1)
pp2 <- eRm::person.parameter(object2)
theta1 <- unlist(pp1$thetapar)*rnorm(nrow(it.SPADI2))
theta2 <- unlist(pp2$thetapar)
threshtable1 <- eRm::thresholds(object1)$threshtable[[1]]
threshtable2 <- eRm::thresholds(object2)$threshtable[[1]]
objlist <- list(beta = list(beta1 = beta1, beta2 = beta2),
                theta = list(theta1 = theta1, theta2 = theta2),
                threshtable = list(threshtable1 = threshtable1,
                                   threshtable2 = threshtable2))
p <- ggImap(objlist, legend.title = "Gender", legend.labels = c("Male", "Female"))
ggPImap(p, theta = list(theta1 = theta1, theta2 = theta2),
             legend.title = "Gender", legend.labels = c("1", "2"))
}
}
\references{
Christensen, K. B., Makransky, G. and Horton, M. (2017)
\emph{Critical Values for Yen’s Q3: Identification of Local Dependence in the Rasch Model Using Residual Correlations, Applied Psychological Measurement, Vol. 41(3), 178-194},\cr
\doi{https://doi.org/10.1177/0146621616677520}.\cr
Yen W. M. (1984)
\emph{Effects of local item dependence on the fit and equating performance of the three-parameter logistic model, Applied Psychological Measurement, Vol. 8, 125-145},\cr
\doi{10.18637/jss.v039.i05}.
}
