% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggQ3star.R
\name{ggQ3star}
\alias{ggQ3star}
\title{Heatmap plot function for Yen's Q3}
\usage{
ggQ3star(
  object,
  method = c("circle", "square"),
  type = c("lower", "upper", "full"),
  labels = c(NA, "upper", "lower", "full", "none"),
  markQ3star = c("circle", "square", "pch", NA),
  ggtheme = theme_minimal,
  title,
  show.legend = TRUE,
  legend.title,
  show.diag,
  colors,
  outline.color,
  outline.width = 0,
  lab_col = "black",
  lab_size = 4,
  pch = 8,
  pch.col = "black",
  pch.cex = 2,
  tl.cex = 12,
  tl.col = "black",
  tl.srt = 45,
  digits = 2
)
}
\arguments{
\item{object}{\code{Q3} object, typically result of \link[RASCHplot]{Q3}.}

\item{method}{character, the visualization method of Q3 matrix to be
used. Allowed values are "circle" (default), "square".}

\item{type}{character, "lower" (default), "upper", or "full" display.}

\item{labels}{character, "upper" (default), "lower", "full" or "none"/NULL, Q3 value labels.}

\item{markQ3star}{character, for highlighting the \eqn{Q_{3\ast}}{Q3*} value by a rectangle.}

\item{ggtheme}{ggplot2 function or theme object. Default value is \code{theme_minimal}. Allowed values are the official ggplot2 themes including \code{theme_gray}, \code{theme_bw}, \code{theme_classic}, \code{theme_void}, .... Theme objects are also allowed (e.g., \code{theme_classic()}).}

\item{title}{character, title of the graph.}

\item{show.legend}{logical, if TRUE (default) the legend is displayed.}

\item{legend.title}{a character string for the legend title.}

\item{show.diag}{NULL or logical, whether display the correlation
coefficients on the principal diagonal. If \code{NULL} (default), the default is to
show diagonal correlation for \code{type = "full"} and to remove it when
\code{type} is one of "upper" or "lower".}

\item{colors}{a vector of 3 colors for low, mid and high correlation values.}

\item{outline.color}{the outline color of square or circle. Default value is "#CFCFC2".}

\item{outline.width}{the outline width of square or circle. Default value is 0 (no outline).}

\item{lab_col, lab_size}{size and color to be used for the correlation
coefficient labels when showed.}

\item{pch}{add character on the glyphs of Q3star value (only valid when markQ3star is TRUE). Default value is 4.}

\item{pch.col, pch.cex}{the color and the cex (size) of pch (only valid when
markQ3star is TRUE).}

\item{tl.cex, tl.col, tl.srt}{the size, the color and the string rotation of
text label (variable names).}

\item{digits}{Decides the number of decimal digits to be displayed (Default is 2).}
}
\description{
Visualise the matrix of Yen's Q3 statistics for a \code{Q3} object using \link[ggplot2]{ggplot2}, e.g., with mixed circle visualization (upper triangle) and numbers (lower triangle). The \eqn{Q_{3\ast}}{Q3*} value may be highlighted by a rectangle. Great attention has been paid to details and appearance may be further tweaked through the function arguments or the `ggplot2` functionality.
}
\examples{
# Load data
data(SPADI)
# Remove incomplete cases
SPADI.complete <- SPADI[complete.cases(SPADI), ]
# Subset items
it.SPADI <- SPADI.complete[, 9:16]
# Compute Yen's Q3 statistics and the \eqn{Q_{3\ast}}{Q3*} value
q3obj <- Q3(items = it.SPADI, method.item = "CML", method.person = "WML", model = "RMP")
# Visualise Yen's Q3
ggQ3star(object = q3obj)
# The \eqn{Q_{3\ast}}{Q3*} value is obtained for items D4 and D5:
q3obj$Q3star

# With the reasoning that these items hold the same information,
# we create a new variable as the sum of D4 and D5:
it.SPADI$D4D5 <- it.SPADI$D4 + it.SPADI$D5
# and remove the two:
it.SPADI.2 <- it.SPADI[,-c(4,5)]
# We compute new Yen's Q3 statistics:
q3obj2 <- Q3(items = it.SPADI.2, method.item = "CML", method.person = "WML", model = "RMP")
ggQ3star(q3obj2)

}
\references{
Christensen, K. B., Makransky, G. and Horton, M. (2017)
\emph{Critical Values for Yen’s Q3: Identification of Local Dependence in the Rasch Model Using Residual Correlations, Applied Psychological Measurement, Vol. 41(3), 178-194},\cr
\doi{https://doi.org/10.1177/0146621616677520}.\cr
Yen W. M. (1984)
\emph{Effects of local item dependence on the fit and equating performance of the three-parameter logistic model, Applied Psychological Measurement, Vol. 8, 125-145},\cr
\doi{10.18637/jss.v039.i05}.
}
