% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{nobsnitms_confirm}
\alias{nobsnitms_confirm}
\title{Create a modal dialog UI when files of item parameters and person parameters are the same}
\usage{
nobsnitms_confirm()
}
\description{
Create a modal dialog UI when files of item parameters and person parameters are the same
}
