% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PARmat.R
\name{PARmat}
\alias{PARmat}
\title{Parameter vector to parameter matrix function}
\usage{
PARmat(x, par, byrows = c("levels", "items"))
}
\arguments{
\item{x}{Matrix of item responses.}

\item{par}{Vector of item parameters.}

\item{byrows}{Items or response levels in rows.}
}
\value{
Matrix of item parameters.
}
\description{
This function transform a vector of item parameters from a polytomous Rasch model to a matrix of item parameters where rows correspond to item categories and columns correspond to items or vice versa.
}
\details{
...
}
\examples{
library(eRm)
data(SPADI)
SPADI.complete <- SPADI[complete.cases(SPADI), ]
it.SPADI <- SPADI.complete[, 9:16]
model.SPADI <- eRm::PCM(it.SPADI)
thr <- PARmat(x = it.SPADI, par = model.SPADI$betapar)

}
