% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMstats-utils.R
\name{pcmfct}
\alias{pcmfct}
\title{PCM function}
\usage{
pcmfct(theta, b, ii)
}
\arguments{
\item{theta}{Vector of person parameters.}

\item{b}{Vector or matrix of item parameters.}

\item{ii}{item index}
}
\description{
PCM function
}
