% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcMfits.R
\name{pcMfits}
\alias{pcMfits}
\title{Model fit for polytomous Rasch model}
\usage{
pcMfits(
  method.item = c("PCML", "CML", "JML", "MML"),
  method.person = c("WML", "MLE"),
  dat
)
}
\arguments{
\item{method.item}{Estimation method for item parameters.}

\item{method.person}{Estimation method for person parameters.}

\item{dat}{Matrix of item responses.}
}
\value{
An object with S3 class \code{"RASCHfits"}.  \item{call}{the call that produced this
object}
}
\description{
Model fit for polytomous Rasch model
}
