% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Q3.R
\name{plot.Q3}
\alias{plot.Q3}
\title{Heatmap plot function for Yen's Q3}
\usage{
\method{plot}{Q3}(
  x,
  markQ3star = TRUE,
  lower.col = NULL,
  upper.col = NULL,
  col.lim = NULL,
  tl.pos = c("lt", "ld", "td", "d", "n"),
  tl.cex = 0.8,
  tl.col = "black",
  number.cex = 0.8,
  colQ3star = "black",
  ...
)
}
\arguments{
\item{x}{\code{Q3} object, typically result of \link[RASCHplot]{Q3}.}

\item{markQ3star}{Logical flag for highlighting the \eqn{Q_{3\ast}}{Q3*} value by a rectangle.}

\item{lower.col}{Passed as col parameter to the lower matrix.}

\item{upper.col}{Passed as col parameter to the upper matrix.}

\item{col.lim}{= NULL,}

\item{tl.pos}{Character or logical, position of text labels. If character, it must be one of 'lt', 'ld', 'td', 'd' or 'n'. 'lt' (default) means left and top, 'ld' means left and diagonal, 'td' means top and diagonal(near), 'l' means left, 'd' means diagonal, 'n' means don't add text-label.}

\item{tl.cex}{Numeric, for the size of text label (variable names).}

\item{tl.col}{The color of text label.}

\item{number.cex}{The cex parameter to send to the call to text when writing the correlation coefficients into the plot.}

\item{colQ3star}{Color of rectangle highlighting the \eqn{Q_{3\ast}}{Q3*} value.}

\item{...}{Other paramters to be passed through to plotting functions. See \link[corrplot]{corrplot.mixed} for further details.}
}
\description{
Visualise the matrix of Yen's Q3 statistics for a \code{Q3} object, e.g., with mixed circle visualization (upper triangle) and numbers (lower triangle). The \eqn{Q_{3\ast}}{Q3*} value may be highlighted by a rectangle. Great attention has been paid to details and appearance may be further tweaked through the function arguments.
}
\examples{
data(SPADI)
SPADI.complete <- SPADI[complete.cases(SPADI), ]
it.SPADI <- SPADI.complete[, 9:16]
q3obj <- Q3(items = it.SPADI, method.item = "CML", method.person = "WML", model = "RMP")
plot(q3obj)
it.SPADI$D4D5 <- it.SPADI$D4 + it.SPADI$D5
it.SPADI.2 <- it.SPADI[,-c(4,5)]
q3obj2 <- Q3(items = it.SPADI.2, method.item = "CML", method.person = "WML", model = "RMP")
plot(q3obj2)

}
\references{
Christensen, K. B., Makransky, G. and Horton, M. (2017)
\emph{Critical Values for Yen’s Q3: Identification of Local Dependence in the Rasch Model Using Residual Correlations, Applied Psychological Measurement, Vol. 41(3), 178-194},\cr
\doi{https://doi.org/10.1177/0146621616677520}.\cr
Yen W. M. (1984)
\emph{Effects of local item dependence on the fit and equating performance of the three-parameter logistic model, Applied Psychological Measurement, Vol. 8, 125-145},\cr
\doi{10.18637/jss.v039.i05}.
}
