% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Q3star.R
\name{plot.Q3star}
\alias{plot.Q3star}
\title{Heatmap plot function}
\usage{
\method{plot}{Q3star}(x, ...)
}
\arguments{
\item{x}{\code{Q3star} object, typically result of \link[RASCHplot]{Q3star}.}

\item{...}{Other paramters to be passed through to plotting functions.}
}
\description{
Visualise the Yen's Q3 statistics for a \code{Q3star} object.
}
