% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.RASCHq3.R
\name{plot.RASCHq3}
\alias{plot.RASCHq3}
\title{Density plot function}
\usage{
\method{plot}{RASCHq3}(
  x,
  extreme = c("max", "min", "star"),
  probs,
  breaks,
  labels,
  colours,
  xtitle,
  title,
  col.outline = 0,
  alpha.ribbon = 1,
  title.legend = "Quantiles",
  ...
)
}
\arguments{
\item{x}{Vector of statistics to be plotted.}

\item{extreme}{Values to be plotted}

\item{probs}{Vector of percentiles.}

\item{breaks}{Vector of breaks.}

\item{labels}{Vector of labels}

\item{colours}{Vector of colours}

\item{xtitle}{Character string for x axis title}

\item{title}{Character string for plot title}

\item{col.outline}{Colour of outline of density}

\item{alpha.ribbon}{Transparency (alpha) of density.}

\item{title.legend}{Legend title.}

\item{...}{Other paramters to be passed through to plotting functions.}
}
\description{
Visualise the empirical distribution of Yen's Q3 for a \code{RASCHq3} object.
}
