% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.RASCHstats.R
\name{plot.RASCHstats}
\alias{plot.RASCHstats}
\title{Density plot function}
\usage{
\method{plot}{RASCHstats}(
  x,
  type = c("Outfit", "Infit", "FitResid"),
  extreme = c("min", "max"),
  probs = NULL,
  breaks = NULL,
  labels = NULL,
  colours = NULL,
  xtitle = NULL,
  title = NULL,
  col.outline = 0,
  alpha.ribbon = 1,
  title.legend = "Quantiles",
  ...
)
}
\arguments{
\item{x}{Vector of statistics to be plotted.}

\item{type}{Type of item fit statistic.}

\item{extreme}{Values to be plotted}

\item{probs}{Vector of percentiles.}

\item{breaks}{Vector of breaks.}

\item{labels}{Vector of labels}

\item{colours}{Vector of colours}

\item{xtitle}{Character string for x axis title}

\item{title}{Character string for plot title}

\item{col.outline}{Colour of outline of density}

\item{alpha.ribbon}{Transparency (alpha) of density.}

\item{title.legend}{Legend title.}

\item{...}{Other paramters to be passed through to plotting functions.}
}
\description{
Visualise the empirical distribution of an item fit statistic of choice for a \code{RASCHstats} object.
}
