% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPImap.R
\name{print.ggPImap}
\alias{print.ggPImap}
\title{Print a ggPImap object}
\usage{
\method{print}{ggPImap}(x, newpage = grDevices::dev.interactive(), ...)
}
\arguments{
\item{x}{ggPImap object.}

\item{newpage}{Should a new page (i.e., an empty page) be drawn before the
ggPImap is drawn?}

\item{...}{ignored}
}
\description{
\code{ggPImap} objects are created from \code{ggMarginal}. This is the S3
generic print method to print the result of the scatterplot with its marginal
plots.
}
\seealso{
\code{\link{ggMarginal}}
}
\keyword{internal}
