% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPImap_hist.R
\name{print.ggPImap_hist}
\alias{print.ggPImap_hist}
\title{Print a ggPImap_hist object}
\usage{
\method{print}{ggPImap_hist}(x, newpage = grDevices::dev.interactive(), ...)
}
\arguments{
\item{x}{ggPImap_hist object.}

\item{newpage}{Should a new page (i.e., an empty page) be drawn before the
ggPImap is drawn?}

\item{...}{ignored}
}
\description{
\code{ggPImap_hist} objects are created from \code{ggPImap_hist}. This is the S3
generic print method to print the result of the scatterplot with its marginal
plots.
}
\seealso{
\code{\link{ggPImap}}
}
\keyword{internal}
