% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMstats-utils.R
\name{print.simRMstats}
\alias{print.simRMstats}
\title{Print a simRMstats object}
\usage{
\method{print}{simRMstats}(x, ...)
}
\arguments{
\item{x}{\code{simRMstats} object}

\item{\dots}{additional print arguments}
}
\description{
Print a summary of the simRMstats.
}
\seealso{
\code{plot} methods.
}
