% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q3.R
\name{Q3}
\alias{Q3}
\title{Compute Yen's Q3 from item fit}
\usage{
Q3(object, ...)
}
\arguments{
\item{object}{Object of class \code{RASCHresiduals}.}

\item{...}{Other arguments.}
}
\description{
Compute Yen's Q3 from item fit
}
