% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainyRASCH.R
\name{rainyRASCH}
\alias{rainyRASCH}
\title{Histogram for Person-Item Map}
\usage{
rainyRASCH(
  theta,
  delta,
  rain.title,
  cloud.title,
  rain.colour,
  rain.fill,
  cloud.colour,
  cloud.fill,
  drop.size
)
}
\arguments{
\item{theta}{The person parameters to use for creating the marginal plots.}

\item{delta}{Input delta (item-category threshold) parameters.
This must be a matrix or list of matrices in case of stratification.}

\item{rain.title}{Label of axis for the latent dimension.}

\item{cloud.title}{Label of axis for the person parameter distribution.}

\item{rain.colour}{The color of the outline of the points.}

\item{rain.fill}{The color of the fill of the points.}

\item{cloud.colour}{The color of outline of the density.}

\item{cloud.fill}{The color of fill of the density.}

\item{drop.size}{The size of the points relative to the binwidth.}
}
\description{
Visualise the alignment of the item and person parameters.
}
\examples{
data(SPADI)
SPADI.complete <- SPADI[complete.cases(SPADI), ]
it.SPADI <- SPADI.complete[, 9:16]
object <- eRm::PCM(it.SPADI)
delta <- beta2delta(beta = object$betapar, x = it.SPADI)
pp <- eRm::person.parameter(object)
theta <- unlist(pp$thetapar)
pp <- eRm::person.parameter(object)

rainyRASCH(theta, delta)


}
