% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RASCHfits.R
\name{RASCHfits}
\alias{RASCHfits}
\title{Model fit for Rasch model}
\usage{
RASCHfits(
  method.item = c("PCML", "CML", "JML", "MML"),
  method.person = c("WML", "MLE"),
  dat
)
}
\arguments{
\item{method.item}{Estimation method for item parameters.}

\item{method.person}{Estimation method for person parameters.}

\item{dat}{Matrix of item responses.}
}
\description{
Model fit for Rasch model
}
