% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RASCHplotpackage.R
\docType{package}
\name{RASCHplot}
\alias{RASCHplot}
\title{\code{RASCHplot} package}
\description{
Visualisation of Rasch item fit
}
\author{
Ann-Sophie Buchardt, Karl Bang Christensen and Sidsel Normann Jensen\cr Maintainer:
Ann-Sophie Buchardt <asbu@sund.ku.dk>
}
