% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RASCHq3.R
\name{RASCHq3}
\alias{RASCHq3}
\title{Compute Yen's Q3}
\usage{
RASCHq3(beta, theta, dat)
}
\arguments{
\item{beta}{Vector or matrix of item parameters.}

\item{theta}{Vector of person parameters.}

\item{dat}{Matrix with item responses.}
}
\description{
Compute Yen's Q3
}
