% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RASCHstats.R
\name{RASCHstats}
\alias{RASCHstats}
\title{Compute fit statistics}
\usage{
RASCHstats(beta, theta, dat)

RASCHstats(beta, theta, dat)
}
\arguments{
\item{beta}{Vector or matrix of item parameters.}

\item{theta}{Vector of person parameters.}

\item{dat}{Matrix with item responses.}
}
\description{
Compute fit statistics

Compute fit statistics
}
