% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RESTSCOREplot.R
\name{RESTSCOREplot}
\alias{RESTSCOREplot}
\title{Item restscore plot}
\usage{
RESTSCOREplot(
  model,
  color.expected = "grey",
  color.observed = "orange",
  size = 4,
  linewidth = 1
)
}
\arguments{
\item{model}{Either a model object or a list of model objects of class \code{Rm} or \code{eRm} returned from the functions \code{RM()}, \code{PCM()}, or \code{RSM()} from the \code{eRm} package.}

\item{color.expected}{Colour of expected restscore point and confidence interval.}

\item{color.observed}{Colour of observed restscore point.}

\item{size}{Colour of points.}

\item{linewidth}{Width of confidence interval.}
}
\value{
Item restscore plot
}
\description{
This function constructs visualization of the observed Gamma coefficient between the score of a single item and the total score of the remaining items compared with the corresponding expected Gamma coefficient under the Rasch model. These plots can be used to investigate item misfit and the association between item scores and restscores under the Rasch model.
}
\details{
...
}
\examples{
library(eRm)
data(SPADI)
SPADI.comp <- SPADI[complete.cases(SPADI[, 4:8]),]
it.SPADI.pain <- SPADI.comp[, 4:8]

model.SPADI.pain <- eRm::PCM(it.SPADI.pain)
strat.vars <- list(gender = SPADI.comp[, "gender"])

br <- cut(rowSums(it.SPADI.pain), quantile(rowSums(it.SPADI.pain), seq(0, 1, 0.25)))
table(br)
strat.vars <- list(over60 = SPADI.comp[, "over60"])
lower.groups <- c(0, 10.5, 16, 19, 25)
DIFplot(model = model.SPADI.pain,
        which.item = 1,
        strat.vars = strat.vars,
        lower.groups = lower.groups)

RESTSCOREplot(model.SPADI.pain)

it.SPADI.pain.0 <- SPADI.comp[SPADI.comp$over60 == 0, 4:8]
it.SPADI.pain.1 <- SPADI.comp[SPADI.comp$over60 == 1, 4:8]

model.SPADI.pain.0 <- eRm::PCM(it.SPADI.pain.0)
model.SPADI.pain.1 <- eRm::PCM(it.SPADI.pain.1)

irt.list <- list("Under 60" = model.SPADI.pain.0,
                 "Over 60" = model.SPADI.pain.1)

RESTSCOREplot(irt.list)

}
