% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testserver.R
\name{RMDserver}
\alias{RMDserver}
\title{Server function for dichotomous App}
\usage{
RMDserver(input, output, session)
}
\arguments{
\item{input}{input set by Shiny.}

\item{output}{output set by Shiny.}

\item{session}{session set by Shiny.}
}
\description{
Server function for dichotomous App
}
