% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMDstats.R
\name{RMDstats}
\alias{RMDstats}
\title{Compute fit statistics for dichotomous Rasch model}
\usage{
RMDstats(delta, theta, dat)
}
\arguments{
\item{delta}{Vector or matrix of item parameters.}

\item{theta}{Vector of person parameters.}

\item{dat}{Matrix with item responses.}
}
\description{
Compute fit statistics for dichotomous Rasch model
}
