% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMfits.R
\name{rMfits}
\alias{rMfits}
\title{Model fit for dichotomous Rasch model}
\usage{
rMfits(
  method.item = c("PCML", "CML", "JML", "MML"),
  method.person = c("WML", "MLE"),
  dat
)
}
\arguments{
\item{method.item}{Estimation method for item parameters.}

\item{method.person}{Estimation method for person parameters.}

\item{dat}{Matrix of item responses.}
}
\description{
Model fit for dichotomous Rasch model
}
