% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMPresiduals.R
\name{RMPresiduals}
\alias{RMPresiduals}
\title{Compute residuals from item fit for polytomous items}
\usage{
RMPresiduals(beta, theta, dat)
}
\arguments{
\item{beta}{Vector or matrix of item parameters.}

\item{theta}{Vector of person parameters.}

\item{dat}{Matrix with item responses.}
}
\description{
Compute residuals from item fit for polytomous items
}
