% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMPserver.R
\name{RMPserver}
\alias{RMPserver}
\title{Server function for polytomous App}
\usage{
RMPserver(input, output, session)
}
\arguments{
\item{input}{input set by Shiny.}

\item{output}{output set by Shiny.}

\item{session}{session set by Shiny.}
}
\description{
Server function for polytomous App
}
