% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMPstats.R
\name{RMPstats}
\alias{RMPstats}
\title{Compute fit statistics for polytomous Rasch model}
\usage{
RMPstats(delta, theta, dat)
}
\arguments{
\item{delta}{Matrix K columns of item-category threshold parameters for K items with at most M categories and number of rows equal to (maximum) number of response categories and NA assigned to empty categories.}

\item{theta}{Vector of person parameters.}

\item{dat}{Matrix with item responses.}
}
\description{
Compute fit statistics for polytomous Rasch model
}
