% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrasch.R
\name{rrasch}
\alias{rrasch}
\title{Simulate Data under a Rasch model}
\usage{
rrasch(model = c("RMD", "RMP"), theta, beta, B, M = NULL, mi = NULL)
}
\arguments{
\item{model}{Character. "RMD" for dichotomous Rasch model. "RMP" for partial credit model.}

\item{theta}{Input person parameters.}

\item{beta}{Input item parameters. If items are dichotomous \code{beta} should be a vector. If items a polytomous \code{beta} should be a matrix or data.frame with number of columns equal to number of items and number of rows equal to (maximum) number of response categories and NA assigned to empty categories.}

\item{B}{Number of simulations.}

\item{M}{Maximum (over all items) number of categories. Only if `model = "RMP"`.}

\item{mi}{Vector of maximum number of categories per item. Only if `model = "RMP"`.}
}
\description{
Repeated random sampling to obtain item responses from estimated item and person parameters.
}
