% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRMD.R
\name{rRMD}
\alias{rRMD}
\title{Simulate item responses under a dichotomous Rasch model}
\usage{
rRMD(delta, theta, B)
}
\arguments{
\item{delta}{Input vector of item parameters.}

\item{theta}{Input vector person parameters.}

\item{B}{Number of simulations.}
}
\description{
Repeated random sampling to obtain item responses from estimated item and person parameters.
}
