% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRMPstats.R
\name{rRMPstats}
\alias{rRMPstats}
\title{Sampling distribution of item fit statistics under polytomous Rasch model}
\usage{
rRMPstats(
  delta,
  theta,
  method.item = c("PCML", "CML", "JML", "MML"),
  method.person = c("WML", "MLE"),
  B,
  trace.it = 0
)
}
\arguments{
\item{delta}{Input item-category threshold parameters. Should be a matrix or data.frame with number of columns equal to number of items and number of rows equal to (maximum) number of response categories and NA assigned to empty categories.}

\item{theta}{Input person parameters.}

\item{method.item}{Estimation method for item parameters.}

\item{method.person}{Estimation method for person parameters.}

\item{B}{Number of simulations.}

\item{trace.it}{If \code{trace.it=1}, then progress bars are displayed.}
}
\value{
An object with S3 class \code{"RASCHstats"}.  \item{call}{the call
that produced this object} \item{statobj}{List of item fit statistics.}
\item{method.item}{Estimation method used for item parameters}
\item{method.person}{Estimation method used for person parameters}
}
\description{
Does repeated random sampling to obtain a sampling distribution for an item fit statistic of interest.
}
\details{
For more details see the help vignette:
\code{vignette("simRASCHstats", package = "RASCHplot")}
}
\examples{

data(SPADI)
SPADI.complete <- SPADI[complete.cases(SPADI), ]
it.SPADI <- SPADI.complete[, 9:16]
object <- eRm::PCM(it.SPADI)
pp <- eRm::person.parameter(object)
delta <- beta2delta(beta = object$betapar, x = it.SPADI)
theta <- pp$thetapar$NAgroup1
obj <- rRMPstats(delta, theta, method.item = "CML", method.person = "WML", B = 2)
plot(obj)

}
