% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simRASCHstats.R
\name{simRASCHstats}
\alias{simRASCHstats}
\title{Simulate item responses}
\usage{
simRASCHstats(
  beta,
  theta,
  method.item = c("PCML", "CML", "JML", "MML"),
  method.person = c("WML", "MLE"),
  B,
  model = c("RMD", "RMP"),
  trace.it = 0
)
}
\arguments{
\item{beta}{Input item parameters}

\item{theta}{Input person parameters}

\item{method.item}{Estimation method for item paramters}

\item{method.person}{Estimation method for person parameters}

\item{B}{Number of simulations}

\item{model}{Character string. Either "RMD" for dichotomous items or "RMP" for polytomous items.}

\item{trace.it}{If \code{trace.it=1}, then progress bars are displayed.}
}
\value{
An object with S3 class \code{"RASCHstats"}.  \item{call}{the call
that produced this object} \item{statobj}{List of item fit statistics.}
\item{method.item}{Estimation method used for item parameters}
\item{method.person}{Estimation method used for person parameters}
}
\description{
Does repeated random sampling to obtain a sampling distribution for an item fit statistic of interest.
}
