% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.RASCHstats.R
\name{statistics.RASCHstats}
\alias{statistics.RASCHstats}
\title{Extract the item fit statistics from a RASCHstats object}
\usage{
statistics.RASCHstats(object, ...)
}
\arguments{
\item{object}{RASCHstats object}

\item{\dots}{additional print arguments}
}
\value{
\code{statistics.RASCHstats} produces a list of matrices of fit statistics
}
\description{
A RASCHstats object has the component \code{statobj} which is a list
of item fit statistics returned by the RASCHstats function for each sample.
\code{statistics} method returns a list of matrices of the item fit
statistics. Each row corresponds to an item and each column corresponds to
a sample.
}
\seealso{
\code{RASCHstats}, \code{simRASCHstats}, and \code{plot}
methods.
}
