% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BARplot.R
\name{BARplot}
\alias{BARplot}
\title{Bar plot}
\usage{
BARplot(data, which.item = "all")
}
\arguments{
\item{data}{A model object of class `Rm` or `eRm` returned from the functions `RM()` or `PCM()` from the `eRm` package.}

\item{which.item}{An integer or vector of integers giving the item(s), for which a CICC-plot should be constructed. The default is `which.item = 1`. The argument will not be used if `all.items = TRUE`.}
}
\value{
Bar plot
}
\description{
This function constructs bar plots...
}
\details{
...
}
\examples{
library(iarm)
str(amts)
it.AMTS <- amts[, 4:13]
theme_set(theme_minimal())
BARplot(it.AMTS)
data <- read.table("http://publicifsv.sund.ku.dk/~kach/scaleval/CAT.txt", sep=' ', header = TRUE)
}
