% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CICCplot.R
\name{CICCplot}
\alias{CICCplot}
\title{Conditional Item Characteristic Curves}
\usage{
CICCplot(
  model,
  which.item = 1,
  lower.groups = "all",
  grid.items = FALSE,
  error.bar = TRUE,
  ...
)
}
\arguments{
\item{model}{A model object of class `Rm` or `eRm` returned from the functions `RM()` or `PCM()` from the `eRm` package.}

\item{which.item}{Either an integer or vector of integers giving the item(s), for which a CICC-plot should be constructed. The default is `which.item = 1`. Or a character string `"all"`for constructing CICC plots for all items in the data.}

\item{lower.groups}{A vector for grouping the set of possible total scores into intervals, for which the empirical expected item-score will be calculated and added to the plot. The vector should contain the lower points of the intervals, that the set of possible total scores should be divided into. If zero does not appear in the vector, it will be added automatically. If `lower.groups = "all"` (default), the empirical expected item-score will be plotted for every possible total score.}

\item{grid.items}{Logical flag for arranging the items selected by which.item in grids using the `ggarrange` function from the `ggpubr` package. Default value is `FALSE`.}

\item{error.bar}{Logical flag for adding errorbars illustrating the empirical confidence interval of the observed means of the conditional item score. The confidence intervals are calculated as follows: For each interval l of the total score, induced by the lower-groups argument, the mean x_l, variance var(x_l), and number of observations n_l within the interval of the total score will be calculated. The confidence interval for the mean x_l is then found as \eqn{x_l \pm 2\cdot \sqrt(\frac{var(x_l)}{n_l})}. Default value is `TRUE`.}

\item{...}{Arguments to be passed to `ggarrange`. The arguments will only be used if 'grid.items = TRUE'.}
}
\value{
CICC plot
}
\description{
This function constructs Conditional Item Characteristic Curves for selected items in a Rasch-model. These plots can be used to investigate item misfit.
}
\details{
...
}
\examples{
library(eRm)
data(SPADI)
SPADI.complete <- SPADI[complete.cases(SPADI), ]
it.SPADI <- SPADI.complete[, 9:16]
model.SPADI <- eRm::PCM(it.SPADI)
# Plot item 1
CICCplot(model = model.SPADI)
# Specify lower points of grouping intervals
CICCplot(model = model.SPADI, lower.groups = c(0, 1, 2, 5, 8, 10))
# Plot item 3 and 4
CICCplot(model = model.SPADI, which.item = c(3,4))
# Plot all items
CICCplot(model = model.SPADI, which.item = "all")
# Construc CICC plot for all items and grid the items 2 by 2
#  with a common legend
CICCplot(model = model.SPADI, which.item = "all", grid.items = TRUE,
         common.legend = TRUE, ncol = 2, nrow = 2)
# Change colours and x axis limits for item 3
ciccobj <- CICCplot(model = model.SPADI, which.item = c(3,4))
ciccobj$`3` +
  ggplot2::scale_colour_manual(values = c("burlywood","cornflowerblue")) +
  ggplot2::xlim(c(2,8))

}
