% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CICCplotPCM.R
\name{CICCplot}
\alias{CICCplot}
\title{Conditional Item Characteristic Curves for Rasch models}
\usage{
CICCplot(
  model,
  which.item = 1,
  lower.groups = NULL,
  all.items = F,
  grid.items = T,
  error.bar = F,
  plot.settings = list(color = NULL, plot.title = NULL, x.axis.seq = NULL, y.axis.title
    = NULL)
)
}
\arguments{
\item{model}{A model object returned from either RM() or PCM() from \code{eRm}-package}

\item{which.item}{An integer or vector of integers giving the item(s), for which a CICC-plot should be constructed. 
The default is \code{which.item=1}. The argument will not be used if \code{all.items = TRUE}.}

\item{lower.groups}{A vector used for dividing the set of possible total scores into intervals, for which the empirical expected item-score will be calculated and added to the plot. The vector should contain the lower points of the intervals, that the set of possible total scores should be divided into. If zero does not appear in the vector, it will be added automatically. If \code{lower.groups = "all"}, the empirical expected item-score will be plotted for every possible total score.}

\item{all.items}{Argument of type logical. If \code{TRUE} a CICC-plot will be constructed for all items in the model-object.}

\item{grid.items}{Argument of type logical. If \code{TRUE}, the items selected by \code{which.item} or \code{all.items} will be arranged in grids with at most four plots per grid. Default is \code{grid.items = FALSE}.}

\item{error.bar}{Argument of type logical. If \code{TRUE}, errorbars illustrating the empirical confidence interval of the observed means of the conditional item score will be added to the plot(s). The confidence intervals are calculated as follows: For each interval \eqn{l} of the total score, induced by the lower-groups argument, the mean \eqn{x_l}, variance \eqn{var(x_l)}, and number of observations \eqn{n_l} within the interval of the total score will be calculated. The confidence interval for the mean \eqn{x_l} is then found as \eqn{x_l \pm 2\cdot \sqrt(\frac{var(x_l)}{n_l})}. Default value is \code{error.bar = FALSE}.}

\item{plot.settings}{A list containing settings for the layout of the figure(s):
\itemize{
   \item \code{color}: A list containing the colors which should be used in the plot. The list must be on the form \code{list(expected = "color1", observed = "color2")}. It is possible to choose all colors supported by the \code{ggplot2} package.The default is \code{color = NULL}, which is equivalent to \code{color1} being black and \code{color2} being red.
   \item \code{plot.title}: Argument with the following options for specifying the title of each CICC-plot:
\itemize{
  \item Construct a title with the number of the item(s) with \code{plot.title = "itemnumber"}.
  \item Use the name of the item(s) as the title of each plot using \code{plot.title = "itemnane"}.
  \item Add user-specified titles to each plot by stating all titles in a character vector on the form \code{plot.title = c("Title1", "Title2", ...)}. This option can not be combined with \code{"itemnumber"} and \code{"itemname"}.
  \item To get no title(s) use \code{plot.title = c("", "", ...)}.
}
\item \code{x.axis.seq}: A vector containing the values of the breaks on the x-axis. The default is the complete sequence of all possible total scores. 
\item \code{y.axis.title}: A character, which should be the label on the the y-axis. If no label is given, the default is "Item Score".
}}
}
\value{
A single CICC-plot or a grid of CICC-plots.
}
\description{
This function construct Conditional Item Characteristic Curves for selected items in a Rasch-model. These plots can be used to investigate item misfit.
}
\examples{

}
