% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIFplot.R
\name{DIFplot}
\alias{DIFplot}
\title{Stratified Conditional Item Characteristic Curves}
\usage{
DIFplot(
  model,
  which.item = 1,
  strat.vars = NULL,
  lower.groups = "all",
  grid.items = FALSE,
  error.bar = TRUE,
  ...
)
}
\arguments{
\item{model}{A model object of class `Rm` or `eRm` returned from the functions `RM()` or `PCM()` from the `eRm` package.}

\item{which.item}{Either an integer or vector of integers giving the item(s), for which a CICC-plot should be constructed. The default is `which.item = 1`. Or a character string `"all"`for constructing CICC plots for all items in the data.}

\item{strat.vars}{A named list of categorical variables for stratification.}

\item{lower.groups}{A named list of length `length(strat.vars)` of lists, vectors or a single vector for grouping the set of possible total scores into intervals, for which the empirical expected item-score will be calculated and added to the plot. The vector(s) should contain the lower points of the intervals, that the set of possible total scores should be divided into. If zero does not appear in the vector(s), it will be added automatically. If `lower.groups = "all"` (default), the empirical expected item-score will be plotted for every possible total score. If a list is provided, the arguments should be named corresponding to the `strat.vars`. If the arguments are lists, they should be named corresponding to the levels of the `strat.vars`.}

\item{grid.items}{Logical flag for arranging the items selected by which.item in grids using the `ggarrange` function from the `ggpubr` package. Default value is `FALSE`.}

\item{error.bar}{Logical flag for adding errorbars illustrating the empirical confidence interval of the observed means of the conditional item score. The confidence intervals are calculated as follows: For each interval l of the total score, induced by the lower-groups argument, the mean x_l, variance var(x_l), and number of observations n_l within the interval of the total score will be calculated. The confidence interval for the mean x_l is then found as \eqn{x_l \pm 2\cdot \sqrt(\frac{var(x_l)}{n_l})}. Default value is `TRUE`.}

\item{...}{Arguments to be passed to `ggarrange`. The arguments will only be used if 'grid.items = TRUE'.}
}
\value{
CICC plot
}
\description{
This function constructs Conditional Item Characteristic Curves for selected items in a Rasch-model and allows for stratification on categorical variables. These plots can be used to investigate item misfit.
}
\details{
...
}
\examples{
library(eRm)
data(SPADI)
SPADI.complete <- SPADI[complete.cases(SPADI), ]
it.SPADI <- SPADI.complete[, 9:16]
model.SPADI <- eRm::PCM(it.SPADI)
strat.vars <- list(gender = SPADI.complete[, "gender"])
DIFplot(model = model.SPADI, strat.vars = strat.vars)
strat.vars <- list(gender = SPADI.complete[, "gender"], over60 = SPADI.complete[, "over60"])
DIFplot(model = model.SPADI, strat.vars = strat.vars)
DIFplot(model = model.SPADI, which.item = c(1,2), strat.vars = strat.vars)
lower.groups <- list(gender = c(0, 1, 2, 5, 8, 10), over60 = c(0, 3, 6, 9))
DIFplot(model = model.SPADI, strat.vars = strat.vars, lower.groups = lower.groups)
DIFplot(model = model.SPADI, which.item = c(1,2),
        strat.vars = strat.vars, lower.groups = lower.groups)
lower.groups <- list(gender = list("1" = c(0, 5, 15, 20, 30, 40),
                                   "2" = c(0,  5, 10)),
                     over60 = list("0" = c(0, 10, 15, 20),
                                   "1" = c(0, 8, 15, 17, 18, 22, 29, 38)))
DIFplot(model = model.SPADI, strat.vars = strat.vars, lower.groups = lower.groups)
}
