% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPImap.R
\name{.reconcileScatPlot}
\alias{.reconcileScatPlot}
\title{Internal ggPImap functions}
\usage{
.reconcileScatPlot(p, delta)
}
\arguments{
\item{p}{p}

\item{delta}{delta}
}
\description{
Internal ggPImap functions
}
