% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggQ3star.R
\name{ggQ3star}
\alias{ggQ3star}
\title{Heatmap plot function for Yen's Q3}
\usage{
ggQ3star(
  object,
  method = c("circle", "square"),
  type = c("mixed", "full", "lower", "upper"),
  markQ3star = c("circle", "pch"),
  lower = NULL,
  upper = NULL,
  ggtheme = theme_minimal,
  title = "",
  show.legend = TRUE,
  legend.title = "Q3star",
  show.diag = NULL,
  colors = NULL,
  outline.color = NULL,
  lab = FALSE,
  lab_col = "black",
  lab_size = 4,
  pch = 8,
  pch.col = "black",
  pch.cex = 5,
  tl.cex = 12,
  tl.col = "black",
  tl.srt = 45,
  digits = 2
)
}
\arguments{
\item{object}{\code{Q3star} object, typically result of \link[RASCHplot]{Q3star}.}

\item{method}{character, the visualization method of Q3star matrix to be
used. Allowed values are "circle" (default), "square".}

\item{type}{character, "mixed" (default), "full", "lower" or "upper" display.}

\item{markQ3star}{Logical flag for highlighting the \eqn{Q_{3\ast}}{Q3*} value by a rectangle.}

\item{lower}{character, the visualisation method of lower triangular part of Q3star matrix to be used (only valid when \code{type = "mixed"}). Allowed values are "number" (default), "circle", "square".}

\item{upper}{character, the visualisation method of upper triangular part of Q3star matrix to be used (only valid when \code{type = "mixed"}). Allowed values are "number", "circle" (default), "square".}

\item{ggtheme}{ggplot2 function or theme object. Default value is \code{theme_minimal}. Allowed values are the official ggplot2 themes including \code{theme_gray}, \code{theme_bw}, \code{theme_classic}, \code{theme_void}, .... Theme objects are also allowed (e.g., \code{theme_classic()}).}

\item{title}{character, title of the graph.}

\item{show.legend}{logical, if TRUE (default) the legend is displayed.}

\item{legend.title}{a character string for the legend title.}

\item{show.diag}{NULL or logical, whether display the correlation
coefficients on the principal diagonal. If \code{NULL} (default), the default is to
show diagonal correlation for \code{type = "full"} and to remove it when
\code{type} is one of "upper" or "lower".}

\item{colors}{a vector of 3 colors for low, mid and high correlation values.}

\item{outline.color}{the outline color of square or circle. Default value is "gray".}

\item{lab}{logical value. If TRUE, add correlation coefficient on the plot.}

\item{lab_col, lab_size}{size and color to be used for the correlation
coefficient labels. used when lab = TRUE.}

\item{pch}{add character on the glyphs of Q3star value (only valid when markQ3star is TRUE). Default value is 4.}

\item{pch.col, pch.cex}{the color and the cex (size) of pch (only valid when
markQ3star is TRUE).}

\item{tl.cex, tl.col, tl.srt}{the size, the color and the string rotation of
text label (variable names).}

\item{digits}{Decides the number of decimal digits to be displayed (Default is 2).}
}
\description{
Visualise the Yen's Q3 statistics for a \code{Q3star} object using \link[ggcorrplot]{ggcorrplot} with mixed circle visualization (upper triangle) and numbers (lower triangle). The \eqn{Q_{3\ast}}{Q3*} value may be highlighted by a rectangle.
}
\examples{
data(SPADI)
SPADI.complete <- SPADI[complete.cases(SPADI), ]
it.SPADI <- SPADI.complete[, 9:16]
q3obj <- Q3star(items = it.SPADI, method.item = "CML", method.person = "WML", model = "RMP")
ggQ3star(object = q3obj)
it.SPADI$D4D5 <- it.SPADI$D4 + it.SPADI$D5
it.SPADI.2 <- it.SPADI[,-c(4,5)]
q3obj2 <- Q3star(items = it.SPADI.2, method.item = "CML", method.person = "WML", model = "RMP")
ggQ3star(q3obj2)

}
\references{
Christensen, K. B., Makransky, G. and Horton, M. (2017)
\emph{Critical Values for Yen’s Q3: Identification of Local Dependence in the Rasch Model Using Residual Correlations, Applied Psychological Measurement, Vol. 41(3), 178-194},\cr
\doi{https://doi.org/10.1177/0146621616677520}.\cr
Yen W. M. (1984)
\emph{Effects of local item dependence on the fit and equating performance of the three-parameter logistic model, Applied Psychological Measurement, Vol. 8, 125-145},\cr
\doi{10.18637/jss.v039.i05}.
}
