% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.RASCHq3.R
\name{plot.RASCHq3}
\alias{plot.RASCHq3}
\title{Density plot function}
\usage{
\method{plot}{RASCHq3}(
  x,
  extreme = c("max", "min", "star"),
  probs,
  breaks,
  labels,
  colours,
  xtitle,
  title,
  col.outline = 0,
  alpha.ribbon = 1,
  title.legend = "Quantiles",
  ...
)
}
\arguments{
\item{x}{Vector of statistics to be plotted.}

\item{extreme}{Values to be plotted}

\item{probs}{Vector of percentiles.}

\item{breaks}{Vector of breaks.}

\item{labels}{Vector of labels}

\item{colours}{Vector of colours}

\item{xtitle}{Character string for x axis title}

\item{title}{Character string for plot title}

\item{col.outline}{Colour of outline of density}

\item{alpha.ribbon}{Transparency (alpha) of density.}

\item{title.legend}{Legend title.}

\item{...}{Other paramters to be passed through to plotting functions.}
}
\description{
Visualise the empirical distribution of Yen's Q3 for a \code{RASCHq3} object.
}
\examples{
data(SPADI)
SPADI.complete <- SPADI[complete.cases(SPADI), ]
it.SPADI <- SPADI.complete[, 9:16]
set.seed(1)
object <- eRm::PCM(it.SPADI)
delta <- beta2delta(beta = object$betapar, x = it.SPADI)
pp <- eRm::person.parameter(object)
theta <- pp$thetapar$NAgroup1
sims <- simRASCHq3(delta = delta, theta = theta,
                   method.item = "JML", method.person = "MLE",
                   B = 2, model = "RMP", trace.it = 1)
plot(sims)

}
\references{
Christensen, K. B., Makransky, G. and Horton, M. (2017)
\emph{Critical Values for Yen’s Q3: Identification of Local Dependence in the Rasch Model Using Residual Correlations, Applied Psychological Measurement, Vol. 41(3), 178-194},\cr
\doi{https://doi.org/10.1177/0146621616677520}.\cr
Yen W. M. (1984)
\emph{Effects of local item dependence on the fit and equating performance of the three-parameter logistic model, Applied Psychological Measurement, Vol. 8, 125-145},\cr
\doi{10.18637/jss.v039.i05}.
}
