% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.RASCHstats.R
\name{quantile.RASCHstats}
\alias{quantile.RASCHstats}
\title{Sample Quantiles}
\usage{
\method{quantile}{RASCHstats}(
  x,
  type = c(NULL, "Outfit", "Infit", "FitResid", "tOutfit", "tInfit"),
  extreme = c(NULL, "min", "max"),
  probs
)
}
\arguments{
\item{x}{A \code{RASCHstats} object.}

\item{type}{Type of item fit statistic. Default is NULL (returns quantiles for all types).}

\item{extreme}{Extreme.}

\item{probs}{Vector of percentiles for \code{extreme}.}
}
\description{
Produces sample quantiles corresponding to the given probabilities for a \code{RASCHstats} object. The smallest observation corresponds to a probability of 0 and the largest to a probability of 1.
}
