% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RASCHresiduals.R
\name{RASCHresiduals}
\alias{RASCHresiduals}
\title{Compute residuals from item fit}
\usage{
RASCHresiduals(beta, theta, data)
}
\arguments{
\item{beta}{A numeric vector or matrix of item parameters with compatible dimensions to data.}

\item{theta}{A vector of person parameters with compatible dimensions to data.}

\item{data}{Matrix with item responses.}
}
\description{
Compute residuals from item fit
}
