% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRMP.R
\name{rRMP}
\alias{rRMP}
\title{Simulate item responses under a polytomous Rasch model}
\usage{
rRMP(delta, theta, B)
}
\arguments{
\item{delta}{Matrix (K x (1 + M)) with location (1st column) and K columns of item-category threshold parameters for K items with at most M categories and number of rows equal to (maximum) number of response categories and NA assigned to empty categories.}

\item{theta}{Input person parameters.}

\item{B}{Number of simulations.}
}
\description{
Repeated random sampling to obtain item responses from estimated item and person parameters.
}
\examples{
library(eRm)
data(SPADI)
SPADI.complete <- SPADI[complete.cases(SPADI), ]
it.SPADI <- SPADI.complete[, 9:16]
model.SPADI <- eRm::PCM(it.SPADI)

betamat <- beta2mat(x = it.SPADI, beta = model.SPADI$betapar)
delta <- beta2delta(beta = betamat)
theta <- rnorm(1000)
obj <- rRMP(delta = delta, theta = theta, B = 2)
fit <- eRm::PCM(obj[[1]])
thresholds(fit)

}
